refactor: move EMBEDDING_PROFILES to data-type package
- Move EMBEDDING_PROFILES and EmbeddingProfileId from rag to data-type package
- Remove unused ModelDimensions constant and getModelDimensions function
- Update all imports across the codebase to use data-type package
- Fix client-side build errors caused by Node.js-only dependencies

This change separates embedding profile definitions from server-side
implementation, allowing client components to access profile data
without importing Node.js-specific modules (pg, fs, etc).
