+++ apps/studio.giselles.ai/lib/vector-stores/github/ingest/pull-requests/chunk-store.ts
@@ -1,4 +1,4 @@
-import type { EmbeddingProfileId } from "@giselle-sdk/rag";
+import type { EmbeddingProfileId } from "@giselle-sdk/data-type";
 import { createPostgresChunkStore } from "@giselle-sdk/rag";
 import { getTableName } from "drizzle-orm";
 import z from "zod/v4";
