+++ internal-packages/workflow-designer-ui/package.json
@@ -17,7 +17,6 @@
 		"@giselle-sdk/flow": "workspace:^",
 		"@giselle-sdk/giselle": "workspace:*",
 		"@giselle-sdk/language-model": "workspace:^",
-		"@giselle-sdk/rag": "workspace:*",
 		"@giselle-sdk/text-editor": "workspace:^",
 		"@giselle-sdk/text-editor-utils": "workspace:^",
 		"@giselle/giselle-sdk-tsconfig": "workspace:*",
