+++ packages/rag/src/chunk-store/postgres/postgres-chunk-store.ts
@@ -1,8 +1,8 @@
+import { EMBEDDING_PROFILES } from "@giselle-sdk/data-type";
 import type { z } from "zod/v4";
 import { PoolManager } from "../../database/postgres";
 import { ensurePgVectorTypes } from "../../database/postgres/pgvector-registry";
 import type { DatabaseConfig } from "../../database/types";
-import { EMBEDDING_PROFILES } from "../../embedder/profiles";
 import {
 	ConfigurationError,
 	DatabaseError,
