+++ packages/rag/src/ingest/pipeline.test.ts
@@ -4,8 +4,8 @@ import type { ChunkerFunction } from "../chunker/types";
 import type { DocumentLoader } from "../document-loader/types";
 import { createPipeline } from "./pipeline";
 
-// Mock the embedder profiles module
-vi.mock("../embedder/profiles", () => ({
+// Mock the data-type module for EMBEDDING_PROFILES
+vi.mock("@giselle-sdk/data-type", () => ({
 	EMBEDDING_PROFILES: {
 		1: {
 			provider: "openai",
@@ -26,6 +26,10 @@ vi.mock("../embedder/profiles", () => ({
 			name: "Google gemini-embedding-001",
 		},
 	},
+}));
+
+// Mock the embedder profiles module
+vi.mock("../embedder/profiles", () => ({
 	createEmbedderFromProfile: vi.fn((profileId, apiKey, telemetry) => ({
 		embed: vi.fn(async () => [0.1, 0.2, 0.3]),
 		embedMany: vi.fn(async (texts) => texts.map(() => [0.1, 0.2, 0.3])),
